requirements.txt
# Flask
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-Babel==4.0.0
Flask-Marshmallow

# Database
SQLAlchemy==2.0.36
Marshmallow-SQLAlchemy
psycopg2-binary==2.9.10

# Others
python-dotenv==1.0.1
flasgger==0.9.7.1
Werkzeug==3.1.3
gunicorn==23.0.0
requests==2.32.3
pytest==8.3.4